backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
Werkzeug==2.3.7